Move HeaderFragment to a new file
