+++ AnkiDroid/src/main/java/com/ichi2/anki/Preferences.kt
@@ -48,6 +48,7 @@ import com.ichi2.anki.contextmenu.CardBrowserContextMenu
 import com.ichi2.anki.exception.ConfirmModSchemaException
 import com.ichi2.anki.exception.StorageAccessException
 import com.ichi2.anki.preferences.AboutFragment
+import com.ichi2.anki.preferences.HeaderFragment
 import com.ichi2.anki.preferences.setOnPreferenceChangeListener
 import com.ichi2.anki.provider.CardContentProvider
 import com.ichi2.anki.reviewer.AutomaticAnswerAction
@@ -223,64 +224,6 @@ class Preferences : AnkiActivity() {
     // ----------------------------------------------------------------------------
     // Inner classes
     // ----------------------------------------------------------------------------
-
-    class HeaderFragment : PreferenceFragmentCompat() {
-        override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
-            setPreferencesFromResource(R.xml.preference_headers, rootKey)
-
-            // Reviewing preferences summary
-            findPreference<Preference>(getString(R.string.pref_reviewing_screen_key))!!
-                .summary = buildCategorySummary(
-                getString(R.string.pref_cat_scheduling),
-                getString(R.string.timeout_answer_text)
-            )
-
-            // Sync preferences summary
-            findPreference<Preference>(getString(R.string.pref_sync_screen_key))!!
-                .summary = buildCategorySummary(getString(R.string.sync_account), getString(R.string.automatic_sync_choice))
-
-            // Notifications preferences summary
-            findPreference<Preference>(getString(R.string.pref_notifications_screen_key))!!
-                .summary = buildCategorySummary(
-                getString(R.string.notification_pref_title),
-                getString(R.string.notification_minimum_cards_due_vibrate),
-                getString(R.string.notification_minimum_cards_due_blink),
-            )
-
-            // Accessibility preferences summary
-            findPreference<Preference>(getString(R.string.pref_accessibility_screen_key))!!
-                .summary = buildCategorySummary(
-                getString(R.string.card_zoom),
-                getString(R.string.button_size),
-            )
-
-            // Controls preferences summary
-            findPreference<Preference>(getString(R.string.pref_controls_screen_key))!!
-                .summary = buildCategorySummary(
-                getString(R.string.pref_cat_gestures),
-                getString(R.string.keyboard),
-                getString(R.string.bluetooth)
-            )
-
-            if (isRestrictedLearningDevice) {
-                findPreference<Preference>("pref_screen_advanced")!!.isVisible = false
-            }
-
-            if (DevOptionsFragment.isEnabled(requireContext())) {
-                setDevOptionsVisibility(true)
-            }
-
-            // Set icons colors
-            for (index in 0 until preferenceScreen.preferenceCount) {
-                val preference = preferenceScreen.getPreference(index)
-                preference.icon?.setTint(Themes.getColorFromAttr(requireContext(), R.attr.iconColor))
-            }
-        }
-
-        fun setDevOptionsVisibility(isVisible: Boolean) {
-            findPreference<Preference>(getString(R.string.pref_dev_options_screen_key))!!.isVisible = isVisible
-        }
-    }
 
     abstract class SettingsFragment : PreferenceFragmentCompat() {
         /** @return The XML file which defines the preferences displayed by this PreferenceFragment
